add exclude filter tests
test: RFE: add additional fields for use in audit filter exclude rules
https://github.com/linux-audit/audit-kernel/issues/5

In addition to the originally supported MSGTYPE, test for support for
PID, UID, GID, LOGINUID, (LOGINUID_SET,) subj_user, subj_role,
subj_type, subj_sen, subj_clr and verify that a sampling of others are
not supported.

Since the exclude filter is checked before the exit filter, use a
syscall (open, unlink) filter to verify the messages are blocked.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>