tests_manual: add test for daemon disappeared
This test attempts to provoke a "*NO* daemon at audit_pid=" message
(other than the well-known harmless audit_pid=0 case).  In some cases
this was caused by an rhashtable bug affecting the lookup of the netlink
socket to the daemon.  In other cases, it was thought to be caused by an
exhaustion of the portid corresponding the the usespace process' PID.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>